Ensure that the SDL mouse wheel is updated
This is in line with the equivalent logic in GlfwMouse.

Previously, wheel data was only updated when the SdlMouse `Scroll` event was invoked. Additionally, the scroll wheels data was never reset which lead to incorrect scrolling behavior.
This CL also adds an `IEquatable<ScrollWheel>` implementation to the `ScrollWheel` struct so we can value-compare instances without casting the underlying data to `System.Numerics.Vector2`.