Fix to ImGuiVulkan Experiment to run on MacOSX (#1197)
* Fix to ImGuiVulkan Experiment to run on MacOSX

* Pointing to the correct lib on OSX

The MoltenVK library (dylib) is considered as an ICD (Installable Client Driver) by the Vulkan loader. The Vulkan loader dynamically loads ICDs at execution time. When using the Vulkan loader, ICDs are not linked directly to the application. Similarly, the Vulkan loader dynamically loads the layers at execution time. Layers are not linked directly to the application.
The MoltenVK library is a little unusual because it is built in such a way that an application can link to it directly and run MoltenVK (Vulkan subset) applications without using the loader and layers. But if you want to use the loader and layers, you don't link MoltenVK and the layers to your application. You link only the loader. The loader then dynamically loads the MoltenVK lib as an ICD and dynamically loads the layers if requested.

* Updating ImGuiVulkan project file

Updating to net6.0
Including Constants for different OSes

* Update ImGuiVulkan.csproj

Rolling back OS Constants

* Updating ImGui.Net Package

* Update ImGuiVulkanApplication.cs

Included KHR portability subset to work on MoltenVK.
Bit set into InstanceCreateInfo for Enumerate Portability.

The project works on MacOS 13.1 now on Apple Silicon.

* Fix MacOs Platform Detection

Using a different approach proposed by the corefx team.