Move TransformHandles mod's TypeDiscover into new Scraping namespace
This is so that the ExtractNestedTypes mod can also make use of the logic without having to duplicate code.

This also adds xml doc comments to all public facing members. Most of these new doc comments are inheritdoc tags. The notable doc comments to review are the comment on the class itself and on GetHandleTypes.