Regenerate bindings on Linux
This is because I'm back on Linux for development. Previously there were blockers that were preventing me from working on Linux, but I now have workarounds for these. This does make the commit log messier than I want, so I might squash a few of these at the end to clean up.