Remove constants that are marked as deprecated aliases
These only seem to exist for Vulkan currently.

For Vulkan:
These represent misnamed constants that have now been renamed to be consistent with the rest of the Vulkan spec.

For Silk:
These misnamed constants and their corrected variants cause name collisions when names are prettified (thus leading to compile errors) and cause common prefix determination for enum member names to be more pessimistic (thus leading to longer names).

For users:
Removing these aliases can induce more API breaks, but because these are aliases and always have an alternative, API breaks are simple to fix. Removing these aliases also makes it clearer which variant is the correct variant to use.