Coerce enum backing types to be signed instead of unsigned to ensure consistent cross-platform generation
This generally forces the output on Linux to match the output on Windows (other than the NativeTypeName attribute).
There are some quirks, such as with the VkFlags64 enums. VkFlags64 is typedef'd to uint64_t so the backing type should be ulong, but we change it to long anyway. This arguably isn't correct, but at least leads to consistent generation and simpler code (since we aren't manually excluding enums from being coerced). This also matches Silk 2, where VkFlags64 enums also used long as the backing type.