Prevent enums with negative values from being coerced
Not sure why these were not being coerced when the Vulkan PR was merged, but they started getting coerced in this PR.
This technically is a good thing since it means the coerce backing types code is now working as previously intended.
However, the previously intended implementation missed an edge case, which this commit fixes.