Implement name affix handling
Note that running the generator with this commit unmodified will lead to to a crash in NameUtils.Prettify due to an empty string input.
This is because of VkVendorId. You'll have to replace the exception in NameUtils.Prettify with a return statement (just return `ret`) to have the generator work.